till 13 bids notifications to the client

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -6,6 +6,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -45,13 +46,13 @@ public class AuctionsController(AuctionDbContext context, IMapper mapper,
         return mapper.Map<AuctionDto>(auction);
     }
 
+    [Authorize]
     [HttpPost]
     public async Task<ActionResult<AuctionDto>> CreateAuction([FromBody] CreateAuctionDto auctionDto)
     {
         var auction = mapper.Map<Auction>(auctionDto);
 
-        // TODO: add current user as seller
-        auction.Seller = "test";
+        auction.Seller = User.Identity?.Name ?? throw new InvalidOperationException("User not found");
 
         context.Auctions.Add(auction);
 
@@ -66,6 +67,7 @@ public class AuctionsController(AuctionDbContext context, IMapper mapper,
         return CreatedAtAction(nameof(GetAuction), new { id = auction.Id }, newAuction);
     }
 
+    [Authorize]
     [HttpPut("{id}")]
     public async Task<ActionResult<AuctionDto>> UpdateAuction(Guid id, [FromBody] UpdateAuctionDto auctionDto)
     {
@@ -78,7 +80,7 @@ public class AuctionsController(AuctionDbContext context, IMapper mapper,
             return NotFound();
         }
 
-        // TODO: check seller is the same as current user
+        if (auction.Seller != User.Identity?.Name) return Forbid();
 
         auction.Item.Make = auctionDto.Make ?? auction.Item.Make;
         auction.Item.Model = auctionDto.Model ?? auction.Item.Model;
@@ -95,9 +97,10 @@ public class AuctionsController(AuctionDbContext context, IMapper mapper,
             return BadRequest("Failed to update auction");
         }
 
-        return Ok(mapper.Map<AuctionDto>(auction));
+        return Ok();
     }
 
+    [Authorize]
     [HttpDelete("{id}")]
     public async Task<ActionResult> DeleteAuction(Guid id)
     {
@@ -107,7 +110,7 @@ public class AuctionsController(AuctionDbContext context, IMapper mapper,
             return NotFound();
         }
 
-        // TODO: check seller is the same as current user
+        if (auction.Seller != User.Identity?.Name) return Forbid();
         context.Auctions.Remove(auction);
 
         await publishEndpoint.Publish<AuctionDeleted>(new {Id = auction.Id.ToString()});
